authentication

+++ part4/blog/controller/blog.js
@@ -1,11 +1,15 @@
 const blogRouter = require('express').Router()
 const logger = require('../utils/logger')
 const Blog = require('../models/blog')
+const User = require('../models/user')
+
+const jwt = require('jsonwebtoken')
+const config = require('../utils/config')
+const middleware = require('../utils/middleware')
 
 blogRouter.get('/', async (request, response) => {
-    await Blog.find({}).then(blogs => {
-      response.json(blogs)
-    })
+    const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
+    response.json(blogs)
 })
 
 blogRouter.get('/:id', async(request, response) => {
@@ -16,21 +20,43 @@ blogRouter.get('/:id', async(request, response) => {
     response.status(404).end()}
 })
 
-blogRouter.post('/', async (request, response) => {
+blogRouter.post('/', middleware.userExtractor, async (request, response) => {
   const body = request.body
+  console.log('getting token...')
+  const token = request.token
+  console.log('token gottens')
+  // const decodedToken = jwt.verify(token, config.SECRET)
+  // if (!decodedToken.id) {
+  //   return response.status(401).json({ error: 'token invalid' })
+  // }
+  // const user = await User.findById(decodedToken.id)
+  const user = request.user
   const blog = new Blog({
     title: body.title,
     author: body.important || false,
     url: body.url,
-    likes: body.likes || 0
+    likes: body.likes || 0,
+    user: user.id
   })
   const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
   response.status(201).json(savedBlog)
 })
 
-blogRouter.delete('/:id', async(request, response, next) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+blogRouter.delete('/:id', middleware.userExtractor, async(request, response, next) => {
+  const blog = await Blog.findById(request.params.id)
+  const token = request.token
+  const user = request.user
+  // const decodedToken = jwt.verify(token, config.SECRET)
+  // if (!decodedToken.id) {
+  //   return response.status(401).json({ error: 'token invalid' })
+  // }
+  // const user = await User.findById(decodedToken.id)
+  if ( blog.user.toString() === user.id.toString() ) {
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  }
 })
   
 blogRouter.put('/:id', (request, response, next) => {
@@ -45,9 +71,8 @@ blogRouter.put('/:id', (request, response, next) => {
   
     Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
       .then(updatedBlog => {
-        response.json(updatedBlog)
+        response.status(201).json(updatedBlog)
       })
-      .catch(error => next(error))
   })
 
 module.exports = blogRouter
